remove empty warriors from valid files
